update Rule
xử lí null ở các rule
thay total cho tên của nhóm tổng (vd :^nganhang)

+++ ProjectOOP/src/Model/Process/RuleTopVol.java
@@ -1,5 +1,4 @@
 package ProjectOOP.src.Model.Process;
-import ProjectOOP.src.Model.Handle.DataInput;
 import ProjectOOP.src.Model.Handle.DataOutput;
 import ProjectOOP.src.Model.Handle.SortData;
 
@@ -14,12 +13,12 @@ public class RuleTopVol extends Rules {
         List<DataOutput> sortedListVol = sorting.sort(data, 1);
         for(int i=0;i<3;i++) {
             map.put("vol" + (i + 1), sortedListVol.get(i).getData().getName());
-            map.put("v"+(i+1), Double.toString(sortedListVol.get(i).getData().getVolume()));
+            map.put("v"+(i+1), Integer.toString((int)sortedListVol.get(i).getData().getVolume()));
         }
 
         List<DataOutput> sortedListVolValue = sorting.sort(data, 8);
         map.put("rvol",sortedListVolValue.get(0).getData().getName());
-        map.put("rv",Double.toString(sortedListVolValue.get(0).getData().getVolumeValue()));
+        map.put("rv",Integer.toString((int)sortedListVolValue.get(0).getData().getVolumeValue()));
         return map;
     }
 }
